This is the last code modification before 1.0; only docs to come. We heavily expanded the CTest reach by adding tests for reflow_sentinel and reflow_transmap.  Interestingly, this landed us with ERL-450, an issue with bxor in the face of negative integers.  The bug is fixed, and only remotely of interest; only when the compiler-generated data is too small will it be triggered.  Use erts from 20.1 or 21 to avoid the problem.  See issue #15 for details, and note that CTest reports the problem as follows:
> The following tests FAILED:
> 11 - erlang-test-traffic_light_nightly-reflow_transmap (Failed)
> Errors while running CTest

Resolved issue #14 by setting a proper test for the number
of bits in each place.  This is enforced to assure maximum
deployment of space in integers but, more importantly, to
increase opportunities of sharing resources.

All in all, we now have an extensive test setup that feels
like it won't cause many surprises.